frontend/experiments/archive/requirements.txt
flask==2.2.3
werkzeug==2.2.3
gunicorn==20.1.0
pandas==1.5.3
numpy==1.24.2
plotly==5.13.1
requests==2.28.2
beautifulsoup4==4.11.2
python-dotenv==1.0.0
psycopg2-binary==2.9.5
scipy==1.10.1
python-dateutil==2.8.2